[t132332] Fix error when filtering many2one.
Note this only happened when clicking on the "Search more..." and if the list view in the dialog didnt have display_name as field.
When this was the case, the created dummy model tried to fetch the display name, but failed due to service-call event not triggering up correctly.

This commit fixes this issue.